Minor updates. Removing boxing and unboxing.

+++ src/main/java/jnt/scimark2/CommandLine.java
@@ -44,7 +44,7 @@ public class CommandLine {
             }
 
             if (args.length > current_arg) {
-                benchmarker.setMinimumExperimentTime(Double.valueOf(args[current_arg]).doubleValue());
+                benchmarker.setMinimumExperimentTime(Double.valueOf(args[current_arg]));
             }
         }
